Testing set1

+++ Buyva/app/src/main/java/com/example/buyva/features/cart/cartList/viewmodel/cartViewModel.kt
@@ -24,62 +24,61 @@ import javax.inject.Inject
 class CartViewModel @Inject constructor(
     application: Application,
     private val cartRepo: CartRepo,
-    private val addressRepo: IAddressRepo
+    private val addressRepo: IAddressRepo,
 ) : AndroidViewModel(application) {
+
 
     private val _cartProducts = MutableStateFlow<ResponseState>(ResponseState.Loading)
     val cartProducts: StateFlow<ResponseState> = _cartProducts
-    private val userEmail = FirebaseAuth.getInstance().currentUser?.email ?: ""
-    val cartId = SharedPreferenceImpl.getFromSharedPreferenceInGeneral("CART_ID_${userEmail.lowercase()}")
    private var cartLinesId = MutableStateFlow<List<String>>(emptyList())
     private val _defaultAddress = MutableStateFlow<Address?>(null)
     val defaultAddress: StateFlow<Address?> = _defaultAddress
 
     private val token = SharedPreferenceImpl.getFromSharedPreferenceInGeneral(USER_TOKEN)
-    private val defaultAddressId = SharedPreferenceImpl.getFromSharedPreferenceInGeneral("${DEFAULT_ADDRESS_ID}_$token")
+    private val _updateCartState = MutableStateFlow<ResponseState>(ResponseState.Success(true))
+    val updateCartState: StateFlow<ResponseState> = _updateCartState
 
-
-    fun getCartDetails() {
-        viewModelScope.launch {
-            cartRepo.getCartProducts(cartId ?: "No cart Id Found!").collect { response ->
-                _cartProducts.value = response
-                Log.i("1", "getCartDetails: $response")
-                if (response is ResponseState.Success<*>) {
-                     cartLinesId = ((response.data as GetCartDetailsQuery.Data).cart?.lines?.edges
-                         ?.map { edge ->
-                             edge.node.id
-                         } ?: emptyList()) as MutableStateFlow<List<String>>
-                        }
-                }
-            }
-        }
-
+    private fun getCartId(): String? {
+        val email = FirebaseAuth.getInstance().currentUser?.email?.lowercase() ?: ""
+        return SharedPreferenceImpl.getFromSharedPreferenceInGeneral("CART_ID_$email")
+    }
 
         fun showCart() {
+            val cartId = getCartId()
+            if (cartId.isNullOrBlank()) {  //first time
+                _cartProducts.value = ResponseState.Failure(Throwable("Add first item to the cart!"))
+                return
+            }
             viewModelScope.launch {
                 _cartProducts.value = ResponseState.Loading
                 try {
-                    cartRepo.getCartProductList(cartId ?: "")
-                        .collect { response ->
-                            _cartProducts.value = response
-                        }
+                    if (cartId != null) {
+                        cartRepo.getCartProductList(cartId )
+                            .collect { response ->
+                                _cartProducts.value = response
+                            }
+                    }
                 } catch (e: Exception) {
                     _cartProducts.value = ResponseState.Failure(Throwable(e.message ?: "Unknown"))
                 }
             }
+
         }
 
 
 
     fun removeProductFromCart(productLineId: String) {
+        val cartId = getCartId()
+
         viewModelScope.launch {
-            val cartId = this@CartViewModel.cartId ?: return@launch
-            cartRepo.removeProductFromCart(cartId, productLineId).collect { response ->
-                if (response is ResponseState.Success<*>) {
-                   showCart()
-                    Log.d("1", "Item removed successfully")
-                } else if (response is ResponseState.Failure) {
-                    Log.e("1", "Remove failed: ${response.message}")
+            if (cartId != null) {
+                cartRepo.removeProductFromCart(cartId, productLineId).collect { response ->
+                    if (response is ResponseState.Success<*>) {
+                        showCart()
+                        //  Log.d("1", "Item removed successfully")
+                    } else if (response is ResponseState.Failure) {
+                        //   Log.e("1", "Remove failed: ${response.message}")
+                    }
                 }
             }
         }
@@ -88,24 +87,26 @@ class CartViewModel @Inject constructor(
 
     fun clearCart() {
         viewModelScope.launch {
-            val cartId = this@CartViewModel.cartId ?: return@launch
+            val cartId = getCartId()
 
-            cartRepo.getCartProducts(cartId).collect { response ->
-                if (response is ResponseState.Success<*>) {
-                    val cart = response.data as? GetCartDetailsQuery.Cart
-                    val lineIds = cart?.lines?.edges?.map { it.node.id } ?: emptyList()
+            if (cartId != null) {
+                cartRepo.getCartProducts(cartId).collect { response ->
+                    if (response is ResponseState.Success<*>) {
+                        val cart = response.data as? GetCartDetailsQuery.Cart
+                        val lineIds = cart?.lines?.edges?.map { it.node.id } ?: emptyList()
 
-                    lineIds.forEach { lineId ->
-                        cartRepo.removeProductFromCart(cartId, lineId).collect { removeResponse ->
-                            if (removeResponse is ResponseState.Failure) {
-                                Log.e("1", "Failed to remove item: ${removeResponse.message}")
+                        lineIds.forEach { lineId ->
+                            cartRepo.removeProductFromCart(cartId, lineId).collect { removeResponse ->
+                                if (removeResponse is ResponseState.Failure) {
+                                    //  Log.e("1", "Failed to remove item: ${removeResponse.message}")
+                                }
                             }
                         }
+                        showCart()
+                        //  Log.d("1", "Cart cleared successfully")
+                    } else if (response is ResponseState.Failure) {
+                        // Log.e("1", "Failed to load cart for clearing: ${response.message}")
                     }
-                   showCart()
-                    Log.d("1", "Cart cleared successfully")
-                } else if (response is ResponseState.Failure) {
-                    Log.e("1", "Failed to load cart for clearing: ${response.message}")
                 }
             }
         }
@@ -113,6 +114,8 @@ class CartViewModel @Inject constructor(
 
 
     fun loadDefaultAddress() {
+         val defaultAddressId = SharedPreferenceImpl.getFromSharedPreferenceInGeneral("${DEFAULT_ADDRESS_ID}_$token")
+
         if (token.isNullOrBlank()) return
 
         viewModelScope.launch {
@@ -122,9 +125,33 @@ class CartViewModel @Inject constructor(
                     val default = addressList.find { it.id?.stripTokenFromShopifyGid() == defaultAddressId }
                     _defaultAddress.value = default
                 } else if (response is ResponseState.Failure) {
-                    Log.e("1", "Failed to load addresses: ${response.message}")
+                //    Log.e("1", "Failed to load addresses: ${response.message}")
                 }
             }
         }
     }
+
+    fun updateCartLine( lineId: String, quantity: Int) {
+        val cartId = getCartId()
+
+        viewModelScope.launch {
+            if (cartId != null) {
+                cartRepo.updateCartLine(cartId, lineId, quantity).collect {
+                    when (it) {
+                        is ResponseState.Failure -> {
+                         //   Log.e("3", "Failed to update line: ${it.message.message}")
+                        }
+
+                        is ResponseState.Success<*> -> {
+                            showCart()
+                         //   Log.i("3", "Line updated successfully")
+                        }
+
+                        else -> {}
+                    }
+                }
+            }
+        }
+    }
+
 }
